Implement Process service and Scoping
Implement a very simple process service based on the initial models
created earlier.

Scoping
=======

So far when we dealt with data, we had query expressions that were
translated into JOOQ queries and react components that render that data.
What is missing for a complete application are variables which we will
provide in scopes.

We use a unified context model whose JSON is embedded into the
appropriate context.

( Java side : de.quinscape.exceed.model.context.ContextModel)

Context Model
-------------

The general format for these is

{
    "properties": {
        "increment": {
            "type": "Integer",
            "defaultValue": "{ 2 }"
        }
    },
    "objects": {
        "foo": {
            "type": "Foo"
        }
    },
    "lists": {
        "bars": {
            "query": "query( Foo )"
        }
    }
}

 * "properties" contains a map of property names to property
   definitions with type and default value expression.

 * "objects" defines scoped domain object values of a given domain
   type (WIP)

 * "lists" defines scoped results of a given query expression.
   (WIP)

The scopes within the application share one namespace and thus every
scoped value needs a unique name.

All currently available scopes form a scope chain from which all
expressions can access the scopes with three functions

 * property('increment') Accesses the scoped property 'increment'
 * object('foo') Accesses a scoped domain object 'foo'
 * list('bars') Accesses the scoped domain object list 'barList'

The scope chain currently consists of either
ApplicationContext and SessionContext or ApplicationContext,
SessionContext, ProcessContext depending on whether the current context
is a process view or a standalone view.

We might add a ViewContext if that provdes generally useful.

The "app.json" application model defines "applicationContext" and
"sessionContext" properties and each process model has its own "context"
model.

Session context and application context track changes so that they can
synchronized easily. Session context synchronizes with the current user
servlet session (cluster-friendly write-back-on-change), application
context synchronizes itself with the application DB state via the
application service.

Processes:
=========

Each process has its own directory below "models/process".

    models/process
    `-- test
        |-- test.json
        `-- view
            |-- detail.json
            `-- list.json

Each process directory contains a json file definining the process name
and a view subdirectory containing the process views.

The process views are basically the same as standalone views. Their name
follows the format "process/view" and there are slightly different
component rules. TButton can only be used in process views etc.

Here is a JSON example for a simple counter process:

{
    "startTransition" : {
        "to" : "count-view",
        "action" : "syslog('Entered Test-Process')"
    },
    "context" : {
        "properties" : {
            "value": {
                "type": "Integer",
                "defaultValue": "{ 0 }"
            }
        }
    },
    "states" : {
        "count-view" : {
            "type" : "process.ViewState",
            "transitions" : {
                "inc" : {
                    "to" : "count-view",
                    "action" : "property('value') = property('value') + 1"
                }
            }
	}
    }
}

Every process has exactly one start state. That start state has one exit
transition that is defined by the "startTransition" property.

Every transition can have three properties:

 * "to" defines the name of the target state (required)
 * "action" defines an action expression.
 * "discard" Marks the transition as discarding input data, not
   requiring view-validation to happen ( a "cancel" transition)

The model is available as de.quinscape.exceed.model.process.Transition
and also provides a "from" and the AST of the action expression.

The defined process context defines an integer property with the
default value 0.

The "states" property contains a map of process state names to process
state definitions.

The current process state types are view states and decision states.

(Processes are available as de.quinscape.exceed.model.process.Process on
the java side)

View-States (type = process.ViewState)
--------------------------------------

    {
        "type": "process.ViewState",
        "transitions": {
            "detail": {
                "to": "detail",
                "action": "…"
            },
            "inc": {
                "to": "list",
                "action": "…"
            }
        }
    }

View states handle user interaction by displaying a view with a scoped
context chain containing a process contex. Each view state must have a
"states" entry in the process model and a separate JSON view definition
in the view/ subdirectory.

The "transition" property of a view state contains a map mapping
transition names to transition models.

Decision-States ( type = process.DecisionState)
-----------------------------------------------

Decision states are handled on the server side and execute conditional
actions or route the process flow.

    {
        "type": "process.DecisionState",
        "decisions": [
            {
                "expression": "property('foo') == 1",
                "transition": {
                    "to": "state-a"
                }
            }, {
                "expression": "property('foo') == 2",
                "transition": {
                    "to": "state-b"
                }
            }
        ],
        "defaultTransition": {
            "to": "state-not-12"
        }
    }

The "decisions" property contains a list of decision models with an
"expression" and a "transition" property. If the expression evaluates to
true, the transition is executed.

If none of the expressions is true, the transition defined by the
"defaultTransition" property is executed.

Process-Mapping
===============

Each process requires a special route mapping to work:

    {
        "rootNode": {
            "kids": [
                …
                {
                    "name": "process",
                    "kids" : [{
                        "name" : "{stateId}",
                        "mapping": {
                            "processName": "test"
                        }
                    }]
                }
                …
            ]
        }
    }

Here we see an extract of a routing.json. Apart from the "processName"
property pointing to the mapped process, there also needs to be a
{stateId} variable mapping.

The start state for the process is entered when the browser initially
requests the route mapping without {stateId}. A state id is created and
a redirection to the route mapping with the {stateId} is created. Every
following state transition happens by POST-ing to the same route mapping
URL with a special "_trans" HTTP parameter containing a transition name.
Each POST is redirected to the a new stateId and thus can be reloaded.

process fixup