.navigateTo(): Add urlProvider
We need to be able to set the new URI to something resulting from the
data.

Most urgent use case being setting the new process state URL.

We add a callback function as 3rd argument that receives the view data
block we received via AJAX and tells us what URL we associate that data
with.