Process: Avoid AJAX redirects
We want to redirect the first request following the Redirect-After-POST
pattern, but we don't need to do so for our AJAX requests.

This requires a bit of restructuring. We already enhanced navigateTo to
allow us to build the new URL.

This requires ammending RuntimeContext with a method to set a single location
variable (better than just writing the getLocationParams() map).

We use it to make sure the new state id of the resumed process
execution is used within the system sans redirect.

We only redirect to the new state if it's not an AJAX request.

We draw out the "stateId" name into a constant.