DataList: Add copy constructor
This adds a cheap copy constructor version to the client side data list.

We can use immutable updates on the rows itself fine, but it is better
if we also copy the data-list itself.

Now except for its rows property, the data list should be constant.

The copy construtor variant is used when the first argument of the
constructor is a DataList instance as opposed to RAW datalist json.

The copy constructor will do a shallow copy of the data-list and just
copy all references to fields, including propertyLookup which will
be not recalculated.

It is assumed that the type information with a datalist never changes
and that changes to the rows field happen immutable.

Test copy ctor.