Enums: Move clientside enum handling to domain svc
Adding the enums used in a data-list to the data-list works, but it is
kind of hackish and leads to bad side-effects.

Better to assume that the system has a fixed set of enums which are
known from the start. The enum models are of course also grounds for hot
reloading, but we will for now just make enums be transferred with the
initial page-load and then assumed to be unchanging.

We add a EnumValuesProvider that returns a map of all enums known in the
application domain service if the current request is not an AJAX
request.

We create a new client-side domain service which we move the preexisting
(but unused) client-side domain factory into. The domain service also
keeps track of the known enums,

We modify main.js to initialize the domain service known enums on window
load.

We refactor data-list.js and property-converter.js to receive their
enum knowledge from the domain service and not a data-list.

We also remove the enums from the serverside data list handling