Assignment: Support scoped value field assignment
So far our syntactic sugar for assigments did only support direct
assignment to a scoped object.

	object('MyObject') = ...

Now we can also use assignment to set properties of our objects

	object('MyObject').myField = ...

We add a paht property to SetActionModel which contains the target path
if such a target path exists. We use JSONPathUtil from svenson to do the
write access with JSON naming conventions.

We adjust existing tests and add new ones for the feature.

We allow property() and list() path assignments because it might be
helpful, not because it makes a lot of sense currently.