Process CRUD
We make TButton and the transition it triggers automatically contain
information about the current domain context. This can be either for
referencing purposes but also for editing objects.

When TButton executes and has a DataCursor Context it checks whether the
cursor is pointing to a property and pops if necessary. Then it extracts
the domain object from that point.

In many situations, the type of the domain object to extract will be
unambiguous, but if the cursor points e.g. to the DataList root of a
joined query, the type is no longer unambiguous and needs to be provided
by the domainType prop. If the domainType prop is not given, the
implementation will throw an error if the type is ambiguous.

The extracted domain object is send as POST data with our transition and
is converted in to a Java Domain Object on the other side. The object
will be available during the processing of the transition and in the
next view as "object('current')", the current domain object context.

The domain object context received has a _type field identifying the
domain type within the app and the an id field for its identity. It can
however only contain some of the fields of the object of that type and
that identity. So that object is read fully from the domain service and
the partial transition context is merged into the full object state
before exposing the merged object as object('current').

This and other scoped objects can be used as data attribute for the Form
component as well as parameter for the store() action or for the set
action.

Referencing a scoped value inside JavaScript or an Expression will be
automatically detected by the usage-tracking and will lead to the server
providing information about the scoped variables in the the view.

Updates then happen when client side expressions referring to scoped
values are fed back to the server side via actions / transitions.