VisibleIf: Add general visibleIf support
Add general support for visibleIf expressions on components. When the
component desciptor must define a signal class "visible-if".

In this case the view-renderer will render the expression in the form

!!(expr) && React.createElement(...)

shortcutting the whole component to a "false" when the expression is not
true.

Component Class Refactoring
---------------------------

We already had one boolean setting in the component desciptor and with
visibleIf there would have been a second and so on.

We also have a class system where we already use signal classes and
using these whereever possible seems preferable over more fields and
constructor args.

We refactor old modelAware prop to a class "model-aware" and will keep
using the class system for more boolean settings.

For documentation and tracking purposes we create a ComponentClasses
class and component-classes.js module that contains special signal
classes like "model-aware" and "visible-if".

The module is generated from the java class automatically via a new
tooling job GenerateComponentClasses. We draw out some common
functionality into a JavaSourceUtil