DataList: Change format
The DataList format as we had it was more complicated than it really
needed to be while not really covering all use-cases.

Globally known Domain Types
===========================

The idea that the data-list can have local domain type definitions
seemed attractive at first but now seems as both a source of confusion
and also caused problems.

We reduce the DataList raw format to just have two properties:

 * columns:  a map of column names to domain properties
 * rows:     the data rows as before.

We skip the indirection of columns being just a column descriptor
pointing to a type in the "types" array and instead have full domain
property definitions which we amend by a domainType property pointing
to the name the property belongs to.

To interpret a DataList fully, you now have to know about the domain
types defined in the application. The columns map contains a basically
redundant map of Domain property definitions with a domainType that will
point to the right system type if there is one. There also can be
independent properties that are not part of a domain type. We will use
that to make the overall data scope a DataList with a single row.

For complex properties, the columns may also contain "List" or "Map" or
"DomainType" typed properties that reference domain types from the
known application types that are now outside of the data list scope. So
the columns entry is somewhat redundant (to avoid partial property
definitions mixed with fully defined) but not totally sufficient to
interpret for deeper structures.

This change again transfers to the client side data-list.js
implementation.

We also move out the DataListCursor type into its own module and add an
util module for common functionality. We adjust the data list tests.