AttributeValue: Allow expression only values
We have used the component model AttributeValue class in multiple
locations for its ability to store parsing error states in the model
etc.

This led to the situation where some expression had to be given in {}
and some not without much separation.

We add a new boolean flag forceExpression to the forValue() method to
force expressions without wrapping {} in some places.

Expression { wrapping }
=======================

This leaves us with the following much clearer separation:

 * View models must continue to wrap their expressions in {} as don't
   want to only have expression atttributes.

 * All other expression in prop declarations, contexts and component
   descriptors now work without {}.