Implement Database schema updating
Until now the database just had to match what the app expected, now the
app can create its own database from the models.

Domain Architecture Changes
===========================

We move around things so that we can define different storage types for
our domain types and also which domain types are supposed to be
synchronized with the configured schema service.

Domain type example:

{
    "name": "MyDomainType",
    "storage": "myStorage",
    "properties": [
        {
            "name": "name",
            "type": "PlainText",
            "required": true
        }
    ]
}

A domain type can define a storage property which the name of a spring
bean implementing

    de.quinscape.exceed.runtime.schema.StorageConfiguration

public interface StorageConfiguration
{
    NamingStrategy getNamingStrategy();

    QueryExecutor getQueryExecutor();

    DomainOperations getDomainOperations();

    SchemaService getSchemaService();
}

The new DomainOperations interface does all the basic domain operations
which were formerly done in domain service and or now delegated to
the correct domain operations implementation for the type.

The SchemaService returned is an interface for synchronizing a list of
domain types (those defining a storage config with that schema service).

interface SchemaService
{
    /**
     * Creates or updates the schema according to the given model.
     *
     */
    void synchronizeSchema(RuntimeContext runtimeContext, List<DomainType> value);

    /**
     * Removes the schema associated with the given application model.
     *
     */
    void removeSchema(RuntimeContext runtimeContext);
}

The default updapte algorithm is implemented on top of another interface
named DDLOperations.

There is an implementation for information_schema based databases (only
tested for Postgres for now).

There is also a NoOpSchemaService for storage configurations without
schema updates.