Refactor DataList into DataGraph
Generalize the existing DataList mechanism to a more general DataGraph
format that not only supports (paged) arrays as main object but also
two flavors of objects.

DataGraph Types
===============

The raw JSON format changes a little to express all now existing types
of data graph.

    {
        "type": "ARRAY",
        "colums": {
            "name": {
                "name": "name",
                "type": "PlainText",
                "domainType": "ExampleType"
            }
        },
        "rootObject": [
            {
                "name": "Anna"
            }
        ],
        "rowCount": 1
    }

First of all the graph now has a type field that is either "ARRAY" or
"OBJECT".

"ARRAY" works as before like in the example above. The rows property was
renamed to rootObject.

"OBJECT" data lists come in two flavors. The first is a map like
construct.

    {
        "type": "OBJECT",
        "colums": {
            "*": {
                "type": "PlainText"
            }
        },
        "rootObject": {
            "A": "Anna",
            "B": "Bernd",
            "C": "Claudia"
        },
        "rowCount": 1
    }

A map like data graph has the type "OBJECT" and a single columns entry
with the name "*" meaning that all entries in the map are of the defined
type. (In this example we have a Map<String,String> equivalent)

The other type of "OBJECT" data graph is a structured object where each
name has a different meaning / type.

    {
        "type": "OBJECT",
        "colums": {
            "name": {
                "type": "PlainText"
            },
            "num": {
                "type": "Integer"
            }
        },
        "rootObject": {
            "name": "Anna",
            "num": 21
        },
        "rowCount": 0
    }

Here we have an example of a structured object graph where the root
object has a PlainText name property and an Integer num property.

Both map-like and structured objects support only strings as keys and
arbitrary properties or complex types (maps, lists, domain types)

The rowCount will express the actual number of the root object
properties.

DataGraph Updates
=================

After working with data graphs for the internal editors for a while and
thinking about the way forward in terms of validation and additional
domain modification rules, it became apparent that the way updates
worked so far was suboptimal overall while also being limited.

In generally it seems preferable to think create immutable copies of the
whole DataGraph instance and not just to the root object.

So the change handlers now work on the signature

    (newList, path) => { ... }

where the first argument is a completely new immutably updated DataGraph
instance and the path is the path for the change (unless cursor.update
is used).

Cursor Change Handlers
======================

For the purpose of  validation beyond simple type checking (e.g. "this
name must be unique") and the execution of domain modification rules
(e.g. "If this is changed, that needs to change as well") we introduce
cursor change handlers.

The change handlers are defined at cursor creation time and then
influence the processing and propagation of changes.

When an immutable update happens via a cursor, the first step is to do
the actual immutable update with "react-addons-update" and then create
a new DataGraph instance with the new data.

Then the change handler value of the cursor determines how things
progess:

 * if the change handler is "true" (default) the DataGraph change
   handler is called like before.

 * if the change handler is "false", no change handler will be called.
   Together with the fact that the modifcation functions (push, unshift,
   splice, set, merge, apply and update) now return the copied DataGraph
   this can be used for strictly local cursors that create new copies of
   data graphs without those changes being propagated.

 * if the change handler is function is is called as a change handler
   and the return value determines what happens next.

   If the cursor change handler returns false, the change is canceled
   and not propagated any further (caller is responsible for setting
   errors or similar to make the user know about the cancelation)

   if the cursor change handler returns a DataGraph instance the
   returned instance replaces the initial graph as change and is
   propagated to the DataGraph change handler.

   Every other return value leads to the propagation of the initial data
   graph changes.

The cursor over which the changes happened will reflect the changes,
that is it is switched over to the new list automatically.

Other cursor will keep their initial data graph states and can revert
changes by applying changes to the initial state. Be aware of this issue
and either recreate the cursors or update them to the new list state.

Cursors will now always fetch their value when .get() is called. The
caching of old values seemed like a bad idea and impractical.