Domain Refactoring / system types
Add the ability to declare internal exceed data structures as domain
types.

We introduce /models/domain/system as new subpath in which we parse
DomainTypes, but change the default for storage.

If a domain type in system does not give a storage, the system storage
is used which used system storage operations that are mostly throwing
UnsupportedOperationException (except for create) for now
and a query handler that does no queries.

We might implement those operations later, but for now we're going to
use dedicated component providers to deliver hand-crafted data-packages
for the internal editors. These refer to the system definitions for
editing.

We add models for AppState and AppTranslation both of which preeexisted
as part of the DB dump.

We're getting more and more independent from that dump and might be able
to create an app from zero in the editor soon.

We extend DDL operations to allow for renaming and providing external
access for field data in preparation for domain editing and refactoring.