Implement Translator, translation editor
We replace the dummy translator with a real implementation based on a
TranslationProvider interface providing all translations of the app at
once. The translations are sorted into a map by specificity and kept in
a concurrent map for fast-acess and per-app refresh.

The application model gets an array of locale codes which are normed
locale codes in the format

	"en_EN"

for our purposes. Separator is one underline and there always needs to
be a language and a country defined.

We match the user locale against these supported locales.

 1. Use full match is found
 2. Look for same language with other country
 3. fall back to first locale defined in the supportedLocales

Translation Tag Prefixes
------------------------

We allow the usage of qualifying prefixes within i18n tags in the format

	qualifier:tag

The translation must be provided for the full tag, but the default for
the tag is just  the part after the qualifier.

These allows for both specific definition of translation labels and
shortness of default representation.

The domain objects now generate tags in the form "Type:field" for their
fields so that the default label for fields is just the field name by
default.

We take over the system of OpenSAGA where we have a general i18n
translation map and to allow specialized overriding on both a per
process and a per view basis.

Placeholders
============

We offer the possibility to add placeholders to your translations ( {0},
{1} etc). The placeholders will be replaced by additional arguments to
the i18n function.

The standard would be make the place holders a part of the tag so
it is clear that these args are inserted, so use for example

	i18n("Edit MyObject {0}", name)

to have a translation tag with filled-in placeholder.

Reference Detection
===================

The following sources for translation tags are auto-detected by the
build process.

    * Javascript modules ( our babel-plugin-track-usage)
    * Domain names and qualified field names
    * Component expressions and component attribute default expressions
      (AST inspection)
    * The supportedLocale codes themselves exist as translation. The
      normal recommendation would be to have a locale have its native
      name in all languages

The translation editor will offer a filterable grid to define
translations and a details form that also allows the definition of
special process and view translation rules.