DataCursor: Add ReactLink / ValueLink interop
We can easily implement a requestChange method that executes a $set on
the current cursor path, what we're lacking was the .value field.

Now we had a value and removed with the idea that it is "caching" and
could maybe be optimized later, but there is ReactLink interop is a nice
thing to have an together with the latest changes/clarification in terms
of cursor data flow, the caching / updating seems ok.

The cursor is supposed to be not changing until an update occurs over it
or it is refresh with newList() in which case we produce a new cursor
anway.

So we're not caching .get() in a value field but are providing a
constant value field that we updated after potential changes.

Both of these changes make a cursor fulfill everything that a ReactLink
has / does.

We change the operations of our FormElement Mixin so that it only uses
requestChange for fieldc changes where it used a direct set call. This
way we can also use all of our input fields with a valuelink and a
propertyType prop providing the necessary type information.

This can be quite helpful for the internal editors and might also be
useful in other situations.

We expose two validation functions (columns and change handlers) via the
interceptor mechanism and test them. Restructure the test a bit.