Interceptor testing helper
When testing modules, proxyquire can be really helpful to mock dependencies
for the purpose of the test, so we are using that in any case.

Another pattern that has proven helpful and which is called interceptor
here is a silly little module that can help a lot.

    module.exports = function (data)
    {
        return data || {}
    };

Just that as individual module allows other modules requiring it to
either create internal buffers or expose internal buffers to this module
we can then proxyquire away in our tests.

All while shielding the export from accidental use / change during
normal operation.