Refactor view-service
Refactor the view-service in two ways:

We want the initial view-data to be available to transitive dependencies
of view-service (translations in module-init code etc). For this purpose
we go the unclean way of exposing a global so that the view service can
initialize itself with the view-data before loading the other modules.
This way the initial data is just there for all other purposes, to, and
we don't need to introduce complex promise initialization etc.

The other thing is that we start giving in the map of known components
as prop to <ViewComponent/> so that it is easier to mock components for
the test, without sucky proxyquire setup.

We move the meta data scripts into <head> for out JSP template so that
they're in fact present immediately.