Templates: Implement own primitive template system
So far we've been using JSP mostly because it was there. We don't really
need any of its complexity and suffer from JSP template loading problems
in the way that resource loading is tied to a servlet container and does
not work the way we like to work with resources (easiy overloading in
extensions etc).

Since we moved basically everything UI generation to javascript and the
client side, the needs of the HTML templates are very very minimal.

This commit implements what might be the most simple template engine
possible.

We spit out UTF-8 text like it is after replacing placeholders that
start with a $ and then contain numbers, upper case letters and underline.

We use a single HTML template per application that we use in three
slightly different scenarios:

    * normal application view
    * login view
    * error view

The first two are full js-generated views and need the full setup of
data-injections, view-model etc.

The last one reuses the one template and injects java-generated static
error description markup. It uses no js currently.

Login View
----------

The login view is modeled like all other views but uses special internal
LoginForm component and raw bootstrap field markup to do its thing. It
is a bit special in that is does not need to be mapped as a routing but
instead is invoked by the LoginController when necessary. The login view
and error view don't really happen inside an application, so they need
to know which application design to load etc.

Default Application
-------------------

We introduce the concept of the default application which defaults to
the first application defined in the system in alphabetic application
name order or can define by setting the property "exceed.default-app" in
the spring environment.

Error Handling
--------------

We now do error handling completely ourselves, from translation the
exception to display information about it with the base template.

The view resolver implemention does not actually resolve the view in the
traditional sense. It reuses the same ExceedView instance for all
application views. The view doesn't really know about the error view
being named "error". It instead detects the special error handling by
the presense of the "error" key in the view model.

( See de.quinscape.exceed.runtime.controller.ExceedExceptionResolver and
  de.quinscape.exceed.runtime.config.ExceedViewResolver )

Compression
-----------

We use the ziplet compression filter for our dynamically generated views
but also resources.