Translator: Add namespacing
For now we have generated composite keys with "." for some cases without
attaching any specific meaning to the separator. ( e.g. "Foo.name" for a
domain property "name" in a domain type "Foo")

While it is nice to have this much specificity in our translation tag
space, the keys can get very long and often we don't really want
different translations anyway. How many different ways do you need to
translate a domain property called "name"?

So we introduce namespacing of translation tags. The tag
"namespace:detail" will first look up a translation of the whole tag and
if that is not found also look up just "detail".

In case a translation is not found, only the last part of a namespaced
translation tag is shown.

We detect all namespaced tags after finding i18n references and add
unqualified/namespaced tags for these tags.

We introduce a new ReferenceType QUALIFIER that expresses that lists all
complete keys for the shortened keys.

We add a way for Queries to introduce translation tags references in the
live system

    DataProviderContext#registerTranslations(String)

for simple translation tags and

    DataProviderContext#registerTranslations(DomainType)

to add all translation tags for that domain type.

We add a empty registerTranslations to QueryDataProvider so people can
override that and provide custom translation keys for query.

We add a little help text explaining the reference type filter modes in
a bit more detail.