Components: Autogenerate missing component ids
We still need component models that have a data-provider / get injected
with data to have a component id. Until now we have thrown an error if
there was no id, but we can easily auto-generate an component id
on view model loading.

The only case where we would need a constant component id now is when we
will have components refer to the injections of other components in the
future. But that problem mostly solves itself since you can't reference
what is not there.

We recycle the existing IDCounter to create autogenerate ids in the form
of "id-xxx" where xxx is a minimal base32 representation of a 64 bit
counter.

We make sure no one uses the id reserved for the runtime info block.