Events: Add own crossbrowser event handler
For the absolute majority of cases we're very well covered by react's
crossbrowser event handling, but there are some cases where either react
does not provide what we need (mousewheel) or where we need to add
events outside of markup under react's control. This is for example the
case with drag and drop events where it is better to register mouse move
and mouse up on window for superior usability.

So we add our own minimal event system for these cases. It uses a
feature test approach to detect the environment on the first event and
then remembering that first result forever.

We add special handling for the mouse-wheel event we need.

The module offers a small static API for events:

Event.add(element, eventName, fn, capturing)

Adds an event to the given element. capturing is ignored in IE.

Event.remove(element, eventName, fn, capturing)

Removes an event from the given element. capturing is ignored in IE.

Event.preventDefault(ev)

Prevents the event from executing the default handling.

Event.location(ev, pos)

Updates the given pos.x/pos.y with the position of the given event. We
use the pos reference approach to avoid allocating many many small
objects for mouse movement.