Add CSFR module
We are using SpringSecurity's CSFR (Cross-Site Request Forgery)
protection which uses a special random token to prevent the user from
being tricked into POSTing bad things to an exceed app.

The CSFR-token is embedded in the current view and each POST submit must
contain it in either a special HTTP parameter or HTTP header.

We already expose the CSFR information to the client. Up to now this has
been handled by the AJAX module doing the actual POSTing.

There maybe be however cases in which we want to also access the CSFR
token and parameter for special static forms like the login form.

Now we move the extraction and caching of CSFR information into
the util/csfr.js module.

Methods:

csfr.token()

csfr.tokenHeader()

csfr.tokenParam()