GUI: Implement zoom, refactor d&d
Implement a smooth zooming of the GUIContainer based on the animation of
the viewBox attribute. Use our new event system for mouse-wheel events.

Implement GUIElement drag and drop and GUIContainer dragging based on a
new global-drag helper module that registers one global mousemove and
mouseup handler and routes the events the current active drag object.

The mousedown is still registered on the element itself for precise
location of elements.