Refactor form components
Changes to the general data cursor information flow, expose input field
via the HL component.

We add general ValueLink interoperability in that we only use the simple
.value/requestChange interface for form elements now which still use
immutable .set() internally.

This allows us to use FormElements internally with just a combination of
property type and ValueLink instead of always having to have a full
data graph and cursors.