Introduce View Layouts
So far we've constructed our views just by converting the view model to
a component hierarchy. What is lacking is the ability to have a reusable
layout frame around that. We implement a simple layout system in this
commit.

We create a new top level model de.quinscape.exceed.model.view.Layout
which contains a component model hierarchy starting at the root
property, just like views.

We make application model contain a "defaultLayout" property which
contains the name of the default layout to use. Views can override that
default layout by setting <View layout="myLayout" /> in their view
components. The resulting component hierarchy should offer all things
the normal view components do.

After loading the initial model, we replace the content with the content
of the layout model and replace a special placeholder component <Content/>
in the layout with the view content.

We rename the existing "Layout" model to LayoutMetaData.