XML-Util: Support Layouts
We add a new includeLayout arg to toXml which defaults to false.

In this mode, we skip all elements that are part of the layout model,
that means we render the opening view tag, then search the <Content/>
to only render its children and close the </View>

Completion still needs to be adapted to the new reality of element
identifying paths being wrong now.