React 15.3.0, change build to webpack
Gulp seems to be overly complicated for what we need here and is also
slower to create the final bundle which led to the shimmig of react etc.

This shimming however has again implications when third-party modules
depend on react internals and so draw duplicates into the bundle even
though we shimed it.

Webpack seems to bundle faster or just as fast without shimming, which
makes the setup far more straight-forward

We switch to simple npm run scripts for our build

Build scripts
=============

--------------------------------- snip ---------------------------------

Lifecycle scripts included in exceed-app:
  test
    mocha --compilers js:babel-register -R spec --recursive src/test/js

available via `npm run-script`:
  dist
    NODE_ENV=production webpack -p
  build
    webpack -d
  watch
    webpack -d -w

--------------------------------- snap ---------------------------------

We keep using babel for transpilation and mocha for testing. We set up
babel in the webpack.config.js because we need to deal with optional
plugins etc.

We invoke mocha directly using "babel-register" as compiler to load the
.babelrc we already had.

We need a webpack plugin now to write our detected function calls to
our track-usage.json. We add the plugin we need as example to
babel-plugin-track-usage.

bulk-require
============

Now lacking the bulkify gulp plugin we cann no longer use bulk-require
to auto-load modules from certain folders.

We switch to webpack's require.context feature for that which requires
some changes in action and component service and also tests.

valueLink props are now deprecated in react and we adjust accordingly